Refactor: modularization
