+++ module/role.py
@@ -1,0 +1,75 @@
+from utils import generate_random_color
+import discord
+
+
+async def requires_admin(message, func):
+    if message.author.server_permissions.administrator:
+        return await func(message)
+    return 'コマンドを実行する権限がありません'
+
+
+async def create_role(client, message):
+    arg = message.content.split('/create_role ')[1]
+    if arg.lower() in [role.name.lower() for role in message.server.roles]:
+        return 'その役職は既に存在します'
+    await client.create_role(
+        message.server,
+        name=arg,
+        mentionable=True,
+        color=discord.Colour(generate_random_color()),
+    )
+    return '役職 {} を作成しました'.format(arg)
+
+
+async def delete_role(client, message):
+    arg = message.content.split('/delete_role ')[1].lower()
+    role_names = [role.name.lower() for role in message.server.roles]
+    if arg in role_names:
+        index = role_names.index(arg)
+        role = message.server.roles[index]
+        await client.delete_role(message.server, role)
+        return '役職 {} を削除しました'.format(role.name)
+    return '役職 {} は存在しません'.format(arg)
+
+
+async def set_roles(client, message):
+    add, rm, pd, nt = [], [], [], []
+    role_names = [role.name.lower() for role in message.server.roles]
+    for role_name in message.content.split()[1:]:
+        if role_name.lower() in role_names:
+            index = role_names.index(role_name.lower())
+            role = message.server.roles[index]
+            if role in message.author.roles:
+                rm.append(role)
+            elif role.permissions.administrator:
+                pd.append(role)
+            else:
+                add.append(role)
+        else:
+            nt.append(role_name)
+    msg = ''
+    if add:
+        await client.add_roles(message.author, *add)
+        msg = msg + '\n役職 {} を付与しました'.format(', '.join([r.name for r in add]))
+    if rm:
+        await client.remove_roles(message.author, *rm)
+        msg = msg + '\n役職 {} を解除しました'.format(', '.join([r.name for r in rm]))
+    if pd:
+        msg = msg + '\n役職 {} は追加できません'.format(', '.join([r.name for r in pd]))
+    if nt:
+        msg = msg + '\n役職 {} は存在しません'.format(', '.join(nt))
+    return msg
+
+
+def is_common(role):
+    if role.is_everyone:
+        return False
+    if role.managed:
+        return False
+    if role.permissions.administrator:
+        return False
+    return True
+
+
+def get_role_names(roles, requirements):
+    return [r.name for r in roles if requirements(r)]
