+++ module/utils.py
@@ -1,0 +1,6 @@
+from random import randint
+
+
+def generate_random_color():
+    rgb = [randint(0, 255) for _ in range(3)]
+    return int('0x{:X}{:X}{:X}'.format(*rgb), 16)
